adds methods to predict overextention
This commit adds methods to check for or predict
overextention of the arm. The velocity of the hand is calculated
and then used to predict a future coordinate. That future
coordinate is compared with the extention limits and retures
a value indicating which side of the robot it overextended on,
if it overextended at all.