adds changes after testing Arm PID control
This change fixes various problems encountered while enabling and
testing PID control for Arm joints. Many enable/disable PID commands
were enabling/disabling the wrong thing due to copy/paste errors. The
enable/disable command dependency on the Arm subsystem was removed so
that the commands could be run after running a given PID command to make
it easy to enable/disable individual PID controllers. Names of
TestingDashboard target angles in the ArmDashboardAngleControl command
were corrected. The ArmDashboardAngleControl and ArmOperatorAngleControl
commands had not been registered to TestingDashboard. The
positionConversionFactor changes for Arm joint encoders were removed
because they interfered with the position information retrieved for
angle PID control. A call to set the setpoint of all PID controllers was
added.

PID for the shoulder, elbow, and wrist joints was tested and worked with
the default PID/tolerance values encoded in the Constants class.
The ArmOperatorAngleControl and ArmDashboardAngleControl commands were
tested and operated as expected with this commit.