tunes robot controls and constants
This change reduces the turret degree rotation to 1 from 5 to slow the
turret down in ArmOperatorAngleControl. The change sets the correct
motor rotations per revolution for the turret. The wrist has been
changed to do an incremental change in ArmOperator*Control. The
drive ramp rate has been reduced to 1 from 2 to increase driving
response time.