changes setpoint of elbowPID and adds sequential commands
This commit changes the setpoint of the elbowPID to make it
measured relative to the lower arm, instead of relative to the
ground. This ensures that the forearm should remain parallel to
the lower arm when only the shoulder angle is changed. This commit
also adds two sequential commands, one that is meant to be used
for scoring and the other is meant to be used for recovering from
scoring. However, they will need to be adjusted based on what
the drive team wants.