adds controller changes and modifies ArmToPreset
This commit does a few things. It adjust the controller layouts,
adds boolean parameters to the ArmToPreset command to enable
the code to disable turning to a specific joints (ex only turning
the elbow to a certain angle and not moving any of the other joints.
This also modifies the ArmToHomeState to zero the turret.